+++ library/src/main/java/com/google/android/exoplayer/extractor/ts/AdtsReader.java
@@ -170,10 +170,9 @@ import java.util.Collections;
       Pair<Integer, Integer> audioParams = CodecSpecificDataUtil.parseAacAudioSpecificConfig(
           audioSpecificConfig);
 
-      MediaFormat mediaFormat = MediaFormat.createAudioFormat(MediaFormat.NO_VALUE,
-          MimeTypes.AUDIO_AAC, MediaFormat.NO_VALUE, MediaFormat.NO_VALUE, C.UNKNOWN_TIME_US,
-          audioParams.second, audioParams.first, Collections.singletonList(audioSpecificConfig),
-          null);
+      MediaFormat mediaFormat = MediaFormat.createAudioFormat(null, MimeTypes.AUDIO_AAC,
+          MediaFormat.NO_VALUE, MediaFormat.NO_VALUE, C.UNKNOWN_TIME_US, audioParams.second,
+          audioParams.first, Collections.singletonList(audioSpecificConfig), null);
       frameDurationUs = (C.MICROS_PER_SECOND * 1024L) / mediaFormat.sampleRate;
       output.format(mediaFormat);
       hasOutputFormat = true;