+++ library/src/main/java/com/google/android/exoplayer/extractor/ts/Id3Reader.java
@@ -35,8 +35,8 @@ import com.google.android.exoplayer.util.ParsableByteArray;
 
   public Id3Reader(TrackOutput output) {
     super(output);
-    output.format(MediaFormat.createFormatForMimeType(MediaFormat.NO_VALUE,
-        MimeTypes.APPLICATION_ID3, MediaFormat.NO_VALUE, C.UNKNOWN_TIME_US));
+    output.format(MediaFormat.createFormatForMimeType(null, MimeTypes.APPLICATION_ID3,
+        MediaFormat.NO_VALUE, C.UNKNOWN_TIME_US));
   }
 
   @Override