+++ library/src/main/java/com/google/android/exoplayer/hls/Variant.java
@@ -26,10 +26,14 @@ public final class Variant implements FormatWrapper {
 
   public final String url;
   public final Format format;
+  public final String name;
 
-  public Variant(int index, String url, int bitrate, String codecs, int width, int height) {
+  public Variant(int index, String name, String url, int bitrate, String codecs, int width,
+      int height) {
     this.url = url;
-    format = new Format(Integer.toString(index), MimeTypes.APPLICATION_M3U8, width, height, -1, -1,
+    this.name = name;
+    String formatName = name != null ? name : Integer.toString(index);
+    format = new Format(formatName, MimeTypes.APPLICATION_M3U8, width, height, -1, -1,
         -1, bitrate, null, codecs);
   }