+++ RELEASENOTES.md
@@ -1,10 +1,17 @@
 # Release notes #
 
-### Current dev branch (from r1.5.1) ###
+### Current dev branch (from r1.5.2) ###
 
-* [Nothing yet]
+* Support for FLV (without seeking).
+* MP4: Fix for playback of media containing basic edit lists.
+* Misc bug fixes.
 
-### r1.5.0 ###
+### r1.5.2 ###
+
+* MPEG-TS/HLS: Fix frame drops playing H265 video.
+* SmoothStreaming: Fix parsing of ProtectionHeader.
+
+### r1.5.1 ###
 
 * Enable smooth frame release by default.
 * Added OkHttpDataSource extension.